Update convert_all.ps1
If no MP4 files are found in the running directory, it will switch to recursive mode and search all sub-directories. Once the search is complete, it will ask the user to select a directory and will place all the processed files into the selected folder relative to the sub-folder they were found.